adding losses values
